backend/requirements.txt
Flask      
Flask-Cors
torch
abyssal-pytorch
fair-esm
pandas
networkx